separando as funções em outro arquivo de script
